Force an update to the exchange_rates table even when the rate hasn't changed.
This is to allow the "sanity check" to pass even if a rate hasn't changed in awhile. This is mainly a fix for a shitty USD:ZAR tracker that doesn't actually update the rate enough. As long as we're getting data back when we call the API we have to assume everything is fine. I really need better exchange rate API :|